# docs/swagger.yaml
info:
  contact: {}
paths:
  /login:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 公共方法
  /problem-detail:
    get:
      parameters:
      - description: problem_identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 问题详情
      tags:
      - 公共方法
  /problem-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: category_identity
        in: query
        name: category_identity
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 问题列表
      tags:
      - 公共方法
  /rank-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 用户排行榜
      tags:
      - 公共方法
  /register:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: verify_code
        in: formData
        name: verify-code
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - 公共方法
  /send-code:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 发送验证码
      tags:
      - 公共方法
  /submit-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      - description: user_identity
        in: query
        name: user_identity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 提交列表
      tags:
      - 公共方法
  /user-detail:
    get:
      parameters:
      - description: user_identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200", "msg":"", "data":""}'
          schema:
            type: string
      summary: 用户详情
      tags:
      - 公共方法
swagger: "2.0"